Add files via upload
A value is one of the basic things a program works with, like a letter or a number.  A variable is a name that refers to a value. An assignment statement creates new variables and gives them values. A statement is a unit of code that the Python interpreter can execute.

Exercise 2.3: Write a program to prompt the user for hours and rate per hour to compute gross pay.
Severance, C. (2016). Python for Everybody: Exploring Data Using Python 3. a Creative Commons Attribution.