Add files via upload
A boolean expression is an expression that is either true or false. There are three logical operators: and, or, and not. Conditional statements gives us the ability to write useful programs where conditions can be checked and the behavior of the program changed accordingly.

Exercise 4.6: Rewrite your pay computation with time-and-a-half for overtime and create a function called computepay which takes two parameters (hours and rate).
Severance, C. (2016). Python for Everybody: Exploring Data Using Python 3. a Creative Commons Attribution.