Add files via upload
In the context of programming, a function is a named sequence of statements that performs a computation. When you define a function, you specify the name and the sequence of statements. Later, you can “call” the function by name. The expression in parentheses is called the argument of the function. The argument is a value or variable that we are passing into the function as input. It is common to say that a function “takes” an argument and “returns” a result. The result is called the return value.

Exercise 4.6: Rewrite your pay computation with time-and-a-half for overtime and create a function called computepay which takes two parameters (hours and rate).
Severance, C. (2016). Python for Everybody: Exploring Data Using Python 3. a Creative Commons Attribution.