Add files via upload
A boolean expression is an expression that is either true or false. There are three logical operators: and, or, and not. Conditional statements check conditions and change the behavior of the program accordingly. The boolean expression after the if statement is called the condition. If the logical condition is true, then the indented statement gets executed. If the logical condition is false, the indented statement is skipped.

Exercise 3.1: Rewrite your pay computation with time-and-a-half for overtime and create a function called computepay which takes two parameters (hours and rate).
Severance, C. (2016). Python for Everybody: Exploring Data Using Python 3. a Creative Commons Attribution.