Add files via upload
A while statement:
1. Evaluates the condition, yielding True or False.
2. If the condition is false, it exits the while statement and continues execution at the following statement.
3. If the condition is true, it executes the body and then go back to step 1.
This type of flow is called a loop because the third step loops back around to the top. We call each time we execute the body of the loop an iteration. 

Exercise 1: Write a program that repeatedly reads integers until the user enters “done”. Once “done” is entered, print out (the total, count, and average revised out) the maximum and the minimum of the integers. If the user enters anything other than an integer, detect their mistake using try and except and print an error message and skip to the next integer.
Severance, C. (2016). Python for Everybody: Exploring Data Using Python 3. a Creative Commons Attribution.